+++ pkgs/storage/030-phdf5.sh
@@ -10,13 +10,15 @@ source "bldr.sh"
 # setup pkg definition and resource files
 ####################################################################################################
 
+pkg_ver_list="1.6.10 1.8.2 1.8.8 1.8.9"
+
 pkg_ctry="storage"
 pkg_name="phdf5"
 pkg_vers="1.8.8"
 
-pkg_info="The Parallel-HDF5 technology suite includes HDF5 compiled with MPI-IO to enable parallel file access."
+pkg_info="The Parallel-HDF5 technology suite includes HDF5 compiled with MPI support to enable distributed parallel file access."
 
-pkg_desc="The Parallel-HDF5 technology suite includes HDF5 compiled with MPI-IO to enable parallel file access.
+pkg_desc="The Parallel-HDF5 technology suite includes HDF5 compiled with MPI support to enable distributed parallel file access.
 
 The Parallel-HDF5 technology suite includes:
 
@@ -63,6 +65,7 @@ pkg_cfg="$pkg_cfg:-DHDF5_ENABLE_Z_LIB_SUPPORT=ON"
 pkg_cfg="$pkg_cfg:-DHDF5_ENABLE_ZLIB_SUPPORT=ON"
 pkg_cfg="$pkg_cfg:-DHDF5_ENABLE_SZIP_SUPPORT=ON"
 pkg_cfg="$pkg_cfg:-DHDF5_BUILD_HL_LIB=ON"
+pkg_cfg="$pkg_cfg:-DHDF5_BUILD_CPP_LIB=ON"
 pkg_cfg="$pkg_cfg:-DHDF5_DISABLE_COMPILER_WARNINGS=ON"
 pkg_cfg="$pkg_cfg:-DHDF5_DISABLE_COMPILER_WARNINGS=ON"
 pkg_cfg="$pkg_cfg -DCMAKE_CXX_FLAGS='-I$BLDR_LOCAL_PATH/storage/szip/latest/include -I$BLDR_LOCAL_PATH/storage/szip/latest/src'"
@@ -130,10 +133,13 @@ function bldr_pkg_install_method()
 
     bldr_log_info "Moving to build path: '$BLDR_BUILD_PATH/$pkg_ctry/$pkg_name/$pkg_vers/$make_path' ..."
     bldr_log_split
-    bldr_push_dir "$BLDR_BUILD_PATH/$pkg_ctry/$pkg_name/$pkg_vers/$make_path"
 
-    mv "../release_docs/USING_CMake.txt" "../release_docs/Using_CMake.txt"
-    bldr_log_split
+    bldr_push_dir "$BLDR_BUILD_PATH/$pkg_ctry/$pkg_name/$pkg_vers/$make_path"
+    if [ -f "../release_docs/USING_CMake.txt" ]
+    then
+        mv "../release_docs/USING_CMake.txt" "../release_docs/Using_CMake.txt"
+        bldr_log_split
+    fi
     bldr_pop_dir
 
     # call the standard BLDR compile method
@@ -156,45 +162,56 @@ function bldr_pkg_install_method()
 }
 
 ####################################################################################################
-# build and install pkg as local module
+# build and install each pkg version as local module
 ####################################################################################################
 
-bldr_build_pkg --category    "$pkg_ctry"      \
-               --name        "$pkg_name"    \
-               --version     "$pkg_vers"    \
-               --info        "$pkg_info"    \
-               --description "$pkg_desc"    \
-               --file        "$pkg_file"    \
-               --url         "$pkg_urls"    \
-               --uses        "$pkg_uses"    \
-               --requires    "$pkg_reqs"    \
-               --options     "$pkg_opts"    \
-               --cflags      "$pkg_cflags"  \
-               --ldflags     "$pkg_ldflags" \
-               --config      "$pkg_cfg"
+for hdf5_ver in ${pkg_ver_list}
+do
+    pkg_name="phdf5"
+    pkg_vers="$hdf5_ver"
+    pkg_file="hdf5-$pkg_vers.tar.gz"
+    pkg_urls="http://www.hdfgroup.org/ftp/HDF5/releases/$pkg_name-$pkg_vers/src/$pkg_file"
 
+    #
+    # phdf5 - parallel HDF5 (using MPI collective IO)
+    #
+    pkg_name="phdf5"
+    pkg_cfg="$hdf5_cfg"
 
-####################################################################################################
+    bldr_build_pkg --category    "$pkg_ctry"    \
+                   --name        "$pkg_name"    \
+                   --version     "$pkg_vers"    \
+                   --info        "$pkg_info"    \
+                   --description "$pkg_desc"    \
+                   --file        "$pkg_file"    \
+                   --url         "$pkg_urls"    \
+                   --uses        "$pkg_uses"    \
+                   --requires    "$pkg_reqs"    \
+                   --options     "$pkg_opts"    \
+                   --cflags      "$pkg_cflags"  \
+                   --ldflags     "$pkg_ldflags" \
+                   --config      "$pkg_cfg"
 
-pkg_name="phdf5-16"
-pkg_cfg="$hdf5_cfg:-DH5_USE_16_API=ON"
+    #
+    # phdf5 - parallel HDF5 w/v1.6 legacy API methods
+    #
+    if [[ $(echo $pkg_vers | grep -m1 -c '^1.8' ) > 0 ]]; then
 
-####################################################################################################
-# build and install pkg as local module
-####################################################################################################
+        pkg_name="phdf5-16"
+        pkg_cfg="$hdf5_cfg:-DH5_USE_16_API=ON"
 
-bldr_build_pkg --category    "$pkg_ctry"    \
-               --name        "$pkg_name"    \
-               --version     "$pkg_vers"    \
-               --info        "$pkg_info"    \
-               --description "$pkg_desc"    \
-               --file        "$pkg_file"    \
-               --url         "$pkg_urls"    \
-               --uses        "$pkg_uses"    \
-               --requires    "$pkg_reqs"    \
-               --options     "$pkg_opts"    \
-               --cflags      "$pkg_cflags"  \
-               --ldflags     "$pkg_ldflags" \
-               --config      "$pkg_cfg"
-
-####################################################################################################
+        bldr_build_pkg --category    "$pkg_ctry"    \
+                       --name        "$pkg_name"    \
+                       --version     "$pkg_vers"    \
+                       --info        "$pkg_info"    \
+                       --description "$pkg_desc"    \
+                       --file        "$pkg_file"    \
+                       --url         "$pkg_urls"    \
+                       --uses        "$pkg_uses"    \
+                       --requires    "$pkg_reqs"    \
+                       --options     "$pkg_opts"    \
+                       --cflags      "$pkg_cflags"  \
+                       --ldflags     "$pkg_ldflags" \
+                       --config      "$pkg_cfg"
+    fi
+done
